services/quiz_evaluator_service/requirements.txt
requests
pydantic
python-dotenv
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.13.0